# Enable XTS optimisation for x86_64