### Issues:
Resolves CryptoAlg-929

### Description of changes: 
From the commit history of `aesni-x86_64.pl`, we gather that the xts assembly functions were disabled to save code space. They are re-enabled in this PR.

~This PR is to be rebased on PR #400.~
It was force-pushed on top of PR #445. PR #400 is now in main.

Init and Encrypt:
|      Size   | Before (MB/s)(1) | After (MB/s) | Improvement |
-------------:|-----------------:|-------------:|:------------:
|    16 bytes |       127        |      129     |    1.02x    |
|   256 bytes |       714        |     1473     |    2.06x    |
|  1350 bytes |       915        |     2816     |    3.08x    |
|  8192 bytes |      1006        |     3626     |    3.60x    |
| 16384 bytes |      1010        |     3748     |    3.71x    |

Init and Decrypt:
|      Size   | Before (MB/s)(1) | After (MB/s) | Improvement |
-------------:|-----------------:|-------------:|:------------:
|    16 bytes |       117        |      119     |    1.02x    |
|   256 bytes |       700        |     1373     |    1.96x    |
|  1350 bytes |       915        |     2663     |    2.91x    |
|  8192 bytes |      1004        |     3610     |    3.60x    |
| 16384 bytes |      1007        |     3735     |    3.71x    |

By submitting this pull request, I confirm that my contribution is made under the terms of the Apache 2.0 license.